Adding new tests wating for master to add more
