+++ src/test/java/com/bootcamp/finalProject/ServicesTests/PartServiceImplTests.java
@@ -1,0 +1,19 @@
+package com.bootcamp.finalProject.ServicesTests;
+
+
+import com.bootcamp.finalProject.repositories.PartRepository;
+import com.bootcamp.finalProject.services.PartService;
+import com.bootcamp.finalProject.services.PartServiceImpl;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class PartServiceImplTests {
+
+    @Autowired
+    PartRepository partRepository;
+
+
+    PartServiceImpl partService = new PartServiceImpl(partRepository);
+}
